BUG/MINOR: quic: Possible leak during probing retransmissions
qc_dgrams_retransmit() prepares two list of frames to be retransmitted into
two datagrams. If the first datagram could not be sent, the TX buffer will
be purged with the prepared packet and its frames, but this was not the case for
the second list of frames.

Must be backported in 2.7.

(cherry picked from commit c6bec2a3afdd537a617fbd9e1ae88a50f81f689d)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>