BUG/MINOR: quic: Useless probing retransmission in draining or killing state
The timer task responsible of triggering probing retransmission did not inspect
the state of the connection before doing its job. But there is no need to
probe the peer when the connection is in draining or killing state. About the
draining state, this is even forbidden.

Must be backported to 2.7 and 2.6.

(cherry picked from commit d21c628ffd598913e47ed28230ac670b52479ef6)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>