MINOR: quic: implement oneline format for "show quic"
Add a new output format "oneline" for "show quic" command. This prints
one connection per line with minimal information. The objective is to
have an equivalent of the netstat/ss tools with just enough information
to quickly find connection which are misbehaving.

A legend is printed on the first line to describe the field columns
starting with a dash character.

This should be backported up to 2.7.

(cherry picked from commit 2273af11e003503767075a1f19bf1201ac8f100f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>