BUG/MINOR: mux-quic: do not prevent shutw on error
Since recent modification of MUX error processing, shutw operation was
skipped for a connection reported as on error. However, this can caused
the stream layer to not be notified about error. The impact of this bug
is unknown but it may lead to stream never closed.

To fix this, simply skip over send operations when connection is on
error while keep notifying the stream layer.

This should be backported up to 2.7.

(cherry picked from commit 3fd40935d91e315e95ba03304d46f9dffb3ff395)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>