BUG/MINOR: quic: Possible crash when dumping version information
->others member of tp_version_information structure pointed to a buffer in the
TLS stack used to parse the transport parameters. There is no garantee that this
buffer is available until the connection is released.

Do not dump the available versions selected by the client anymore, but displayed the
chosen one (selected by the client for this connection) and the negotiated one.

Must be backported to 2.7 and 2.6.

(cherry picked from commit b97169629696dd7cc4c8a3f15aa0c960ee113272)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>