DEV: flags/show-sess-to-flags: only retrieve hex digits from hex fields
Some fields are followed by a comma or a closing parenthesis and we
take them because we read everything that's not a space. Better be
stricter, we're causing warnings about incorrect hex format when
they're passed to printf.

(cherry picked from commit fd1047ae6eaf5fba472dfb484e7465438196a93c)
Signed-off-by: Willy Tarreau <w@1wt.eu>