BUILD: mjson: Fix warning about unused variables
clang 15 reports unused variables in src/mjson.c:

  src/mjson.c:196:21: fatal error: expected ';' at end of declaration
    int __maybe_unused n = 0;

and

  src/mjson.c:727:17: fatal error: variable 'n' set but not used [-Wunused-but-set-variable]
    int sign = 1, n = 0;

An issue was created on the project, but it was not fixed for now:

   https://github.com/cesanta/mjson/issues/51

So for now, to fix the build issue, these variables are declared as unused.
Of course, if there is any update on this library, be careful to review this
patch first to be sure it is always required.

This patch should fix the issue #1868. It be backported as far as 2.4.

(cherry picked from commit 16e314150aca771d6509ca79897006f78fd9441c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 744ffd33ea781af4db811633fe09061a89bf012f)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 43cd1c403f843ea2f04014b23e1c2da003c9eed8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>