MINOR: spoe: Don't stop disabled proxies
SPOE register a signal handler to be able to stop SPOE applets ASAP during
soft-stop. Disabled proxies must be ignored at this staged because they are
not fully configured.

For now, it is useless but this change is mandatory to fix a bug.

(cherry picked from commit 7f4ffad46e98b1004a348d4853693035400a0443)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>