BUG/MINOR: mux-quic: handle properly Tx buf exhaustion
Since the following patch
  commit 6c501ed23bea953518059117e7dd19e8d6cb6bd8
  BUG/MINOR: mux-quic: differentiate failure on qc_stream_desc alloc
it is not possible to check if Tx buf allocation failed due to a
configured limit exhaustion or a simple memory failure.

This patch fixes it as the condition was inverted. Indeed, if buf_avail
is null, this means that the limit has been reached. On the contrary
case, this is a real memory alloc failure. This caused the flag
QC_CF_CONN_FULL to not be properly used and may have caused disruption
on transfer with several streams or large data.

This was detected due to an abnormal error QUIC MUX traces. Also change
in consequence trace for limit exhaustion to be more explicit.

This must be backported up to 2.6.

(cherry picked from commit 1611a7659b25a9bdc226790b4acb185eb3de2414)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit a8c004033434e3ffdc8d3f6813290dd51e853544)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>