BUG/MINOR: tcp-rules: Don't shortened the inspect-delay when EOI is set
A regression was introduced with the commit cb59e0bc3 ("BUG/MINOR:
tcp-rules: Stop content rules eval on read error and end-of-input").  We
should not shorten the inspect-delay when the EOI flag is set on the SC.

Idea of the inspect-delay is to wait a TCP rule is matching. It is only
interrupted if an error occurs, on abort or if the peer shuts down. It is
also interrupted if the buffer is full. This last case is a bit ambiguous
and discutable. It could be good to add ACLS, like "wait_complete" and
"wait_full" to do so. But for now, we only remove the test on SC_FL_EOI
flag.

This patch must be backported to all stable versions.

(cherry picked from commit 2d5a5665fe6c03f1fae1af22d21138347d9acab1)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit cfcdeb0417d55a448f4eafe0d4e4801efab46538)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>