BUG/MINOR: quic: Wrong token length check (quic_generate_retry_token())
This bug would never occur because the buffer supplied to quic_generate_retry_token()
to build a Retry token is large enough to embed such a token. Anyway, this patch
fixes quic_generate_retry_token() implementation.

There were two errors: this is the ODCID which is added to the token. Furthermore
the timestamp was not taken into an account.

Must be backported to 2.6 and 2.7.

(cherry picked from commit 6d6ddb2ce58807b8defb0b38d9d524b958278ce4)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>