REGTESTS: log: Reduce again response inspect-delay for last_rule.vtc
It was previously reduced from 10s to 1s but it remains too high, espeically
for the CI. It may be drastically reduced to 100ms. Idea is to just be sure
we will wait for the response before evaluating the TCP rules.

(cherry picked from commit 292619fc90a918342fd552e572139ed3fcc8d2bb)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>