BUG/MINOR: clock: fix the boot time measurement method for 2.6 and older
clock_update_date() changed between 2.6 and 2.7, it used to always
update now_ms prior to 2.7 while it doesn't anymore later if not
relevant. As such, calling it to measure the boot time causes the
backport in commit a2e5d62d58 ("MINOR: clock: measure the total boot
time") to still apply the offset between date and now_ms, in addition
to the extra delay, resulting in the double time for the checks and
still the well-known symptoms of desynched clocks in "show sess all".
Updating "date" only using a regular gettimeofday() is much more
reliable here apparently. This is only for 2.6 and 2.4.